+++ plugins/BEdita/Core/src/Model/Action/SignupUserAction.php
@@ -15,6 +15,7 @@ namespace BEdita\Core\Model\Action;
 
 use BEdita\Core\Model\Entity\AsyncJob;
 use BEdita\Core\Model\Entity\User;
+use BEdita\Core\Model\Validation\CustomUrlValidationProvider;
 use BEdita\Core\Utility\LoggedUser;
 use Cake\I18n\Time;
 use Cake\Mailer\MailerAwareTrait;
@@ -97,50 +98,21 @@ class SignupUserAction extends BaseAction
     protected function validate(array $data)
     {
         $validator = new Validator();
-        $validator->setProvider('signup', $this);
+        $validator->setProvider('customUrl', new CustomUrlValidationProvider());
 
         $validator
             ->requirePresence('activation_url')
             ->add('activation_url', 'customUrl', [
                 'rule' => 'isValidUrl',
-                'provider' => 'signup',
+                'provider' => 'customUrl',
             ])
 
             ->add('redirect_url', 'customUrl', [
                 'rule' => 'isValidUrl',
-                'provider' => 'signup',
+                'provider' => 'customUrl',
             ]);
 
         return $validator->errors($data);
-    }
-
-    /**
-     * Checks that a value is a valid URL or custom url as myapp://
-     *
-     * @param string $value The url to check
-     * @param array $context The validation context
-     * @return bool
-     */
-    public function isValidUrl($value, array $context = [])
-    {
-        // check for a valid scheme (https://, myapp://,...)
-        $regex = '/(?<scheme>^[a-z][a-z0-9+\-.]*:\/\/).*/';
-        if (!preg_match($regex, $value, $matches)) {
-            return false;
-        }
-
-        // if scheme is not an URL protocol then it's a custom url (myapp://) => ok
-        if (!preg_match('/^(https?|ftps?|sftp|file|news|gopher:\/\/)/', $matches['scheme'])) {
-            return true;
-        }
-
-        if (!empty($context['providers']['default'])) {
-            $provider = $context['providers']['default'];
-        } else {
-            $provider = (new Validator())->getProvider('default');
-        }
-
-        return $provider->url($value, true, $context);
     }
 
     /**
