+++ plugins/BEdita/Core/tests/TestCase/Model/Action/SignupUserActionTest.php
@@ -171,19 +171,4 @@ class SignupUserActionTest extends TestCase
             ],
         ]);
     }
-
-    /**
-     * Test `isValidUrl()` not with validation (without context)
-     *
-     * @return void
-     */
-    public function testIsValidUrlWithoutContext()
-    {
-        $action = new SignupUserAction();
-        $result = $action->isValidUrl('https://example.com');
-        $this->assertTrue($result);
-
-        $result = $action->isValidUrl('https://examplecom');
-        $this->assertFalse($result);
-    }
 }
